LPS-131524 explicit casting to Long because HashMapBuilder calls the Supplier method instead of the Long one.

+++ modules/apps/portal-vulcan/portal-vulcan-impl/src/main/java/com/liferay/portal/vulcan/internal/graphql/servlet/GraphQLServletExtender.java
@@ -1673,7 +1673,8 @@ public class GraphQLServletExtender {
 							dataFetchingEnvironment,
 							HashMapBuilder.<String, Serializable>put(
 								"siteId",
-								dataFetchingEnvironment.getArgument("siteId")
+								(Long)dataFetchingEnvironment.getArgument(
+									"siteId")
 							).build()))
 			).build());
